+++ packages/gui/src/components/primitives/ColorPicker/PreviewPalettePicker.tsx
@@ -15,18 +15,21 @@ export function PreviewPalettePicker({
   theme,
 }: MiniPaletteProps) {
   const [isCollapsed, setIsCollapsed] = useState<boolean>(true)
-  const collapsedRows = 2
+  const collapsedRows = 8
   return (
     <>
       <Label sx={{ py: 2 }}>Theme Colors</Label>
-      <div sx={{ display: 'flex', flexDirection: 'column', gap: 1 }}>
+      <div sx={{ '& *': {
+        transition: 'all .2s ease-in-out', 
+      }
+      }}>
         {theme?.colors?.map((colorGroup: any, i: number) => {
           if (isCollapsed && i >= collapsedRows) {
             return null
           }
 
           return (
-            <div key={i} sx={{ display: 'flex', gap: '.125rem' }}>
+            <div key={i} sx={{ display: 'grid', gridTemplateColumns: 'repeat(12,1fr)', gap: '.25rem', mb: 1 }}>
               {colorGroup.colors.map((color: any, i: number) => {
                 const selected = value === color.value
 
@@ -35,14 +38,16 @@ export function PreviewPalettePicker({
                     key={color.id}
                     title={`${colorGroup.name}.${i}`}
                     sx={{
+                      appearance: 'none',
+                      WebkitAppearance: 'none',
                       cursor: 'pointer',
-                      border: selected ? '2px solid' : '1px solid',
+                      borderWidth: selected ? '2px' : '1px',
                       borderColor: selected ? '#ff0000' : 'border',
-                      borderRadius: '.25rem',
+                      borderStyle: 'solid',
+                      borderRadius: '9999px',
                       padding: 0,
                       margin: 0,
-                      width: '1.5rem',
-                      height: '1.5rem',
+                      width: '100%',
                       aspectRatio: '1 / 1',
                       backgroundColor: color.value,
                     }}
@@ -68,7 +73,7 @@ export function PreviewPalettePicker({
         }}
         onClick={() => setIsCollapsed(!isCollapsed)}
       >
-        {`${isCollapsed ? 'View Full Theme' : 'Hide'}`}
+        {`${isCollapsed ? 'View All' : 'Hide'}`}
       </a>
     </>
   )