+++ packages/gui/src/lib/codegen/html.ts
@@ -1,4 +1,16 @@
+import { unified } from 'unified'
+import rehypeStringify from 'rehype-stringify'
 import { HtmlNode } from '../../components/html/types'
+import { editorSchemaToHast } from '../transformers/editor-schema-to-hast'
+import { format } from './format'
+
+export const unstyledHtml = async (node: HtmlNode) => {
+  const root = editorSchemaToHast(node, {
+    removeStyleProperty: true,
+  })
+  const output = unified().use(rehypeStringify).stringify(root)
+  return format('html', output)
+}
 
 export const html = async (node: HtmlNode) => {
   const res = await fetch('https://components.ai/api/v1/gui/export/html', {
@@ -9,6 +21,6 @@ export const html = async (node: HtmlNode) => {
     body: JSON.stringify({ html: node }),
   })
 
-  const docHtml: string = await res.text()
-  return docHtml
+  const html = await res.text()
+  return html
 }