+++ packages/gui/src/lib/transformers/plugins/attributes-to-properties.ts
@@ -1,0 +1,5 @@
+import { renameNodeAttribute } from './rename-node-attribute'
+
+export const attributesToProperties = () => (tree: any) => {
+  return renameNodeAttribute(tree, 'attributes', 'properties')
+}