+++ packages/gui/src/lib/transformers/plugins/properties-to-attributes.ts
@@ -1,33 +1,5 @@
-import { visit } from 'unist-util-visit'
+import { renameNodeAttribute } from './rename-node-attribute'
 
 export const propertiesToAttributes = () => (tree: any) => {
-  visit(tree, 'element', (node) => {
-    if (!node.properties) {
-      return
-    }
-
-    const properties: Record<string, string[]> = node.properties
-    const attributes = Object.entries(properties).reduce(
-      (acc: Record<string, string>, curr: [string, string | string[]]) => {
-        const [attr, value] = curr
-
-        if (Array.isArray(value)) {
-          return {
-            [attr]: value.join(' '),
-            ...acc,
-          }
-        }
-
-        return {
-          [attr]: value as string,
-          ...acc,
-        }
-      },
-      {}
-    )
-
-    node.attributes = attributes
-    delete node.properties
-    delete node.position
-  })
+  return renameNodeAttribute(tree, 'properties', 'attributes')
 }