+++ packages/gui/src/types/modules.d.ts
@@ -1,1 +1,2 @@
 declare module 'culori'
+declare module 'escape-html'